+++ mutations/hidden_display_scan.pl
@@ -1,0 +1,1 @@
+s/if \(!str_ends_with\(\$file, '\.mustache'\) && !str_ends_with\(\$file, '\.tsx'\)\) \{/if (true) {/;