+++ mutations/react_literal_badge_scan.pl
@@ -1,0 +1,1 @@
+s/if \(!str_ends_with\(\$file, '\.tsx'\) && !str_ends_with\(\$file, '\.ts'\)\) \{/if (true) {/;