+++ mutations/react_scan_coverage.pl
@@ -1,0 +1,1 @@
+s/glob\(\$root \. '\/js\/esm\/src\/\*\.tsx'\),/glob(\$root . '\/js\/esm\/src\/*.nothing'),/;